Add better thread management

+++ Server/src/Game/Session/GameSession.h
@@ -41,7 +41,7 @@ private:
 	bool IsWin() const;
 
 public:
-	const int GetReplayCount() const { return m_ReplayerIds.size(); }
+	const int GetReplayCount() const { return (int)m_ReplayerIds.size(); }
 	void AddReplayer(TGATPLAYERID id) { m_ReplayerIds.push_back(id); }
 	void Replay();
 	bool IsReplayer(TGATPLAYERID id) const { return std::find(m_ReplayerIds.begin(), m_ReplayerIds.end(), id) != m_ReplayerIds.end(); }
